requirements.txt
accelerate==0.24.1
faiss-cpu==1.7.4
faiss-gpu
huggingface-hub==0.19.4
InstructorEmbedding==1.0.1
langchain==0.0.340
openai==1.3.5
pypdf==3.17.1
PyPDF2==3.0.1
python-dotenv==1.0.0
safetensors==0.4.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
streamlit==1.28.2
tiktoken==0.5.1
torch==2.1.1
transformers==4.35.2
